backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# CORS Headers
django-cors-headers==4.3.1

# Database (PostgreSQL driver - optional, SQLite is default for development)
# psycopg2-binary==2.9.9  # Uncomment if using PostgreSQL

# API Integration
requests==2.31.0

# Environment Variables
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
flake8==6.1.0
black==23.11.0

# Documentation
drf-spectacular==0.26.5

# Utilities
python-dateutil==2.8.2
Pillow==11.0.0

# Storage
boto3==1.34.34
django-storages==1.14.3
